Adding support for AWS CLI tools
These are needed to copy the data and examples from the icehack S3 bucket to each virtual environment.

An IAM user has been created for this purpose, with no permissions as all the icehack data is publicly accessible. We don't the data here because that would make the image size stupidly big and means the data is tied to the virtual environment release - which would be impractical.

AWS CLI is a python module, installed through Pip.