Adding Nginx
This adds the Nginx webserver, which is configured to reverse proxy Jupyter.

In this image only HTTP is used, however it is possible to use HTTPS instead through external provisioning if desired.

This change removes the need for the Jupyter firewall rule to be used, and is replaced by a role to open ports 80/443.